PR Fix: Bounds not updating on depth change
